+++ src/common/string_util.h
@@ -9,7 +9,7 @@
 #include "../core/core.h"
 
 static const char ESCAPE_PREFIX = '\\';
-static const size_t ESCAPE_SEQUENCE_LENGTH = 2;
+static const uint32_t ESCAPE_SEQUENCE_LENGTH = 2;
 
 // Structure to define an escape mapping
 typedef struct {
@@ -20,7 +20,7 @@ typedef struct {
 // Srtucture to define what escaped characters map to which real characters
 typedef struct {
   const EscapeMappingNode *mappings;
-  size_t count;
+  uint32_t count;
   bool preserve_unknown;
 } EscapeConfig;
 
@@ -50,15 +50,15 @@ static const EscapeConfig DEFAULT_ESCAPE_CONFIG = {
 bool string_clean_escape_sequences(char *input, const EscapeConfig *config);
 
 // Custom function to compare string slice with exact token match
-bool string_slice_equals(const char *str_slice, const size_t str_length,
+bool string_slice_equals(const char *str_slice, const uint32_t str_length,
                          const char *token_str);
 
 // Custom version of strcpn using a callback instead of a list of characters
-size_t strspn_callback(const char *str, bool (*predicate)(char c));
+uint32_t strspn_callback(const char *str, bool (*predicate)(char c));
 
 // Creates a new string and writes it into the buffer
 void strip_newline(const char *restrict str, char *restrict buffer,
-                   size_t buffer_size);
+                   uint32_t buffer_size);
 
 // Modify a string in place to get rid of tralining newlines (\n and \r)
-void strip_trailing_newlines(char *str, const size_t n);
+void strip_trailing_newlines(char *str, const uint32_t n);
